requirements.txt
Django==4.2.14
djangorestframework==3.15.2
requests==2.32.3
sqlparse==0.5.0
